+++ src/Public/Connect-AZ.ps1
@@ -1,0 +1,42 @@
+function Connect-AZ {
+    [CmdletBinding()]
+    param (
+    )
+
+    <#
+    .SYNOPSIS
+    Connects to Microsoft Azure platform.
+
+    .DESCRIPTION
+    Connects to Microsoft Azure platform.
+
+    .INPUTS
+    None. You cannot pipe objects to Add-Extension.
+
+    .OUTPUTS
+    // <OBJECTTYPE>. TBD.
+
+    .EXAMPLE
+    PS> Connect-AZ -Credential $Credential
+
+    .LINK
+    http://github.com/blindzero/Connect-MS365
+
+    #>
+
+    # testing if module is available
+    while (!(Test-MS365Module -Module $ModuleName)) {
+        # and install if not available
+        Install-MS365Module -Module $ModuleName
+    }
+    try {
+        Connect-AzAccount | Out-Null
+    }
+    catch {
+        $ErrorMessage = $_.Exception.Message
+        Write-Error -Message "Could not connect to $ServiceItem.`n$ErrorMessage" -Category ConnectionError
+        Break
+    }
+    # set service name into window title if successfully connected
+    Set-WindowTitle -Service $ServiceItem
+}